fix: Improve sidebar workaround for Linux w/Intel GPU
Previous fixes set an opacity value of 1 which affects the sidebar
styling for Linux users.

The problem of corrupted text in the sidebar seems to be related to
stacking contexts therefore using `position: relative` to create a new
one for affected text elements is an alternative workaround that has no
effect on styling.

Ref: https://github.com/laurent22/joplin/issues/7506
Ref: https://github.com/laurent22/joplin/issues/8000